Use ovn-nbctl --wait=hv sync to get end-to-end port binding time
Currently the scale testing waits port state UP in ovn-nb and then
starts next round of port creation and binding. This is inaccurate,
because when CPU is 100% in test farms it can't reflect the real
time that ovn-controllers spent to complete each round of processing.

With the new feature "--wait=hv" [1], and the follow-up "sync" cmd
support [2], we can now wait for the port bindings to be processed
and reflected on all HVs before starting next round, so that the real
processing time can be reflected and evaluate more accurately the
related scalability improvements.

[1] https://github.com/openvswitch/ovs/commit/fa183acc654f7e5da17cd70fc91d6b5b02782183
[2] https://github.com/openvswitch/ovs/commit/de32cec780e8d2a4358075c2449858247881f266