ci/scale-pure-ovn-hosts: remove unused params
The params: ovs_repo, ovs_branch, configflags are not needed
when running the deploy, as these are only used in the construction
of the docker images.

When deploy is hooked into a CI system to test a specific OVS commit sha,
we will need to add support for passing in the OVS_REPO, OVS_BRANCH, and CONFIG_FLAGS
variables. Till then, let's keep it out to avoid potential confusion.