Adapt shell scripts for running in a non-interactive SSH session
Because we running scripts remotely in a non-interactive SSH session,
the PATH is set to the sshd's default value, that is
/usr/local/bin:/usr/bin. This can be seen by running:

  $ ssh -T rally@<node> 'echo $PATH'

Due to that the scripts cannot find the OVS/OVN daemons or tools like
'ip' which are usually installed in /usr/local/sbin or /usr/sbin.

Set the PATH manually from the scripts to fix it.

Signed-off-by: Jakub Sitnicki <jkbs@redhat.com>