Set up tox for running unit tests
Configure tox so that we can run automated tests for ovn-scale-test
using the tests framework from the Rally fork. To do that we install
Rally in editable mode, because Rally's tests are not distributed in its
package. We also place our tests not in the top level directory to avoid
a package naming conflict with Rally.

Tox configuration file is based on the one from the Rally fork but cut
down to a bare minimum. We will reintroduce the left out test
environments (py34, pep8, etc.) gradually. Test requirements were copied
directly from the Rally fork.