Use _get_ovn_controller() for all ovn-nbctl connections.
The _get_ovn_controller() calls set_daemon_socket() to properly set the
socket files for ovn-nbctl connections. This is required whenever
running ovn-nbctl in daemon mode and must be used whenever an ovn-nbctl
connection is instantiated. Otherwise, when running the NB DB in
clustered mode, if leadership changes happen ovn-nbctl will only try to
connect to the local DB instance.

Signed-off-by: Dumitru Ceara <dceara@redhat.com>