Fix ovn-sbctl calls that might hang in a clustered DB scenario.
ovn_fake_multinode scenarios check if a chassis has registered with the
SB DB by running ovn-sbctl commands on the first central instance.
However, if a leadership change happens and the first central instance
is not SB DB leader anymore, the ovn-sbctl call will hang.

To avoid such issues we add the "--no-leader-only" option to ovn-sbctl
calls in this case.

Signed-off-by: Dumitru Ceara <dceara@redhat.com>